refactor: extract functionality to composable
