+++ apps/web/app/components/TableOfContents/types.ts
@@ -1,0 +1,9 @@
+import type { Block } from '@plentymarkets/shop-api';
+
+export interface FlatBlock {
+  uuid: string;
+  label: string;
+  depth: number;
+  block: Block;
+}
+
